backend/app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Julian Stephens <julian@julianstephens.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.88.0"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
sqlalchemy = "^1.4.45"
psycopg2-binary = "^2.9.5"
alembic = "^1.9.0"
python-dotenv = "^0.21.0"
asyncpg = "^0.27.0"
loguru = "^0.6.0"
pyjwt = {extras = ["crypto"], version = "^2.6.0"}


[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
pyright = "^1.1.284"
isort = "^5.11.2"
flake8 = "^6.0.0"
mypy = "^0.991"
flake8-pyproject = "^1.2.2"

[tool.flake8]
per-file-ignores = [
  "app/*/__init__.py:F401",
  "app/*/__init__.py:W391",
]
max-line-length = 88

[tool.black]
line-length = 88
preview = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
